Transfer MFred tests into test suite
